glib-2: Removing docs option completely for now
Due to the complexity of multiple modules needed for docs and tests
I decided to disable this for the moment. Docs require Python and
gdbus-codegen that we split out. One reason why I decided to disable
it was that in one case this would have created a circular depdendency.

If docs are wanted later we just introduce a glib-2-docs module that
will have Python and gdbus-codegen as hard dependencies and add that
module as optional to glib-2.