lunar: introduce lwtf and lcd functions
These functions have been incredibly helpful to me
in the last years and I think it's time to include
them in the official lunar module.

Yes, there is `lvu what`, but `lwtf` is shorter and
has more style :). Also, it mimics the standard `wtf`
tool.

Yes, there is `lvu cd` but that opens a non-interactive subshell
for each use of the function and also drops PS1 etc.
Using a bash function for that enables us to jump between many
modules keeping the same shell environment.