glibc: Small module cleanup, new security patches.
Removed systemd unit files for the time being because nscd currently
cause issues with add_priv_{user,group} if enabled. Until we have a
mechanism available in lunar to handle that we better leave it
disabled. What happens is that the nscd cache isn't updated quickly
enough after add_priv_{user,group} which cause issues for commands
depending on a user or group existing.

Also removed a few old fixes that no longer seems to be required.