systemd: version bumped to 215
This bump has been long overdue. The compat patch is necessary to maintain
semi compatibility with some applications that yet has adapted to the combined
libsystemd without actually installing the compat libs.

We might have to fix a couple of modules to behave that depend on systemd, but
so far the only one I've found that need a small fix is sane-backends. Eventually
we should be able to drop the compat patch that only install a couple of .pc files.