installwatch: patch to initialized if called uninitialized
ld's initialization behavior for shared objects is a mess.
The order is pretty much arbitrary. In theory one can mark a library as
"initfirst" to avoid other shared objects initializing before,
but that's a *unique* mark and libpthread uses it already.

As a result LD_PRELOAD overlays the libc methods *from start*,
but until libinstallwatch is initialized (_init called),
libs end up calling invalid function pointers if they access an
overlayed libc method.

Frustratingly, the llvm version of the latest rustc tries to access
the number of cores `open("/proc/cpuinfo")` during library initialization
(WHYYY?!) and thus segfaults if ld decides to load it before libinstallwatch.

And all I was trying to do is getting a new firefox version compiled...