make: Determine if things are broken before attempting to build
Someone reported that upgrading libffi made make stop working, on
account of how they'd built it with guile, and guile was trying to use
a shared library with a different version number.

Someone proposed a fix of simply getting rid of guile, but that seems to
be throwing out the baby with the bathwater, so here's my idea: if guile
is present, and asked for, but still broken anyway, build the more
minimal version of make without it so that at least your system has a
working make so you can go back and fix everything else.