lvm2: Fix build for gcc 10.2 on i686
For some reason, gcc 10.2 forgets about the -shared flag if it's too
early on the command line, but only on i686 builds. This ugly hack adds
an extra "-shared" later on so that it remembers what it was supposed to
be doing, so that lvm2 can build in a 32-bit environment.