glib-2: Bump to 2.76.0 and removing the invalid switch.
Our current version of glib-2 to old for the current gobject-introspection which will silently
pull in the version of glib-2 it needs. The problem with that it stomps on the system glib-2.
So a install or recompile of gobject-introspection will fail leaving you with a broken glib-2
and the reason for https://github.com/lunar-linux/moonbase-other/issues/2262.