dracut: Use local XSLT resources instead of trying to download them
The "xsltproc" invocations in the dracut Makefile is trying to use a
XSLT stylesheet downloaded from docbook.sourceforge.net (non-SSL HTTP
even!). But these stylesheets are already available on the system if
asciidoc is installed! So use the local copies instead of trying (and
failing) to download them instead.

Now we can finally have man pages for our dracut installations.