ncurses: BUILD: ensure wide and narrow are built
It seems as of ncurses 6.5, wide character ncurses is now the default,
so instead of asking for it to support Unicode, you now have to
explicitly ask it to fail to support Unicode.

So now the narrow build has to have "--disable-widec" and I suppose the
"--enable-widec" in the wide build is now redundant.