gmp: Patch the configure script to work with gcc 15
The irony is that this particular snippet of C code is meant to trigger
a core dump on gcc 3 on Darwin, so I could probably do away with the
entire thing altogether, but that would be too much work to bother with.

So I'll just fix this one bit of code which is deliberately broken
anyway.