ncurses: Version bumped to 6.5 (#3915)
* ncurses: Version bumped to 6.5

* ncurses: Disable C++ support

* ncurses: symlink libncursesw.so to libncurses.so (HT Linux From Scratch)

* ncurses: BUILD: ensure wide and narrow are built

It seems as of ncurses 6.5, wide character ncurses is now the default,
so instead of asking for it to support Unicode, you now have to
explicitly ask it to fail to support Unicode.

So now the narrow build has to have "--disable-widec" and I suppose the
"--enable-widec" in the wide build is now redundant.