Zstd (#3947)
* zstd: No version bump, just using cmake instead so various *.cmake

files get created and installed for those apps looking for them.
Also added lzma-sdk as an optional_depends.

* zstd: Works better with cmake as a depends.

* zstd: Another rework of this module. As noted in my last comment on this

PR there is an issue with the cmake config modules it installls. Notably
with qt6-tools failing its configuration stage, its an issue with defining
libzstd_shared.

Using zstd make or the meson build avoids this problem because qt6-tools
configuration simply moves on with the compile ignoring the
/usr/lib64/cmake/zstd/*.cmake problem when they are not present.